Renamed packages from "service" to "clientService"
