+++ bot/src/main/java/edu/java/bot/commands/ListCommand.java
@@ -8,7 +8,7 @@ import com.pengrad.telegrambot.request.SendMessage;
 import edu.java.bot.models.GenericResponse;
 import edu.java.bot.models.Link;
 import edu.java.bot.models.ListLinksResponse;
-import edu.java.bot.service.BotService;
+import edu.java.bot.clientService.BotService;
 import java.util.Properties;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
