+++ bot/src/test/java/edu/java/bot/BotTest.java
@@ -10,8 +10,8 @@ import edu.java.bot.commands.StartCommand;
 import edu.java.bot.commands.TrackCommand;
 import edu.java.bot.commands.UntrackCommand;
 import edu.java.bot.commandsHolder.CommandsHolder;
-import edu.java.bot.service.BotService;
-import edu.java.bot.service.DefaultBotService;
+import edu.java.bot.clientService.BotService;
+import edu.java.bot.clientService.DefaultBotService;
 import edu.java.bot.updatesListener.MessageUpdatesListener;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.DisplayName;
