+++ bot/src/test/java/edu/java/bot/clientService/DefaultBotServiceTest.java
@@ -1,4 +1,4 @@
-package edu.java.bot.service;
+package edu.java.bot.clientService;
 
 import edu.java.bot.api.httpClient.ScrapperClient;
 import edu.java.bot.api.dto.request.AddLinkRequest;
