+++ bot/src/test/java/edu/java/bot/commands/TrackCommandTest.java
@@ -5,16 +5,12 @@ import com.pengrad.telegrambot.request.SendMessage;
 import edu.java.bot.api.dto.response.ApiErrorResponse;
 import edu.java.bot.models.AddLinkToDatabaseResponse;
 import edu.java.bot.models.GenericResponse;
-import edu.java.bot.service.BotService;
-import edu.java.bot.service.DefaultBotService;
+import edu.java.bot.clientService.DefaultBotService;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Answers;
 import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
 import java.net.URI;
 import java.util.List;
 import java.util.Properties;
