+++ bot/src/test/java/edu/java/bot/commandsHolder/CommandsHolderTest.java
@@ -6,8 +6,8 @@ import edu.java.bot.commands.ListCommand;
 import edu.java.bot.commands.StartCommand;
 import edu.java.bot.commands.TrackCommand;
 import edu.java.bot.commands.UntrackCommand;
-import edu.java.bot.service.BotService;
-import edu.java.bot.service.DefaultBotService;
+import edu.java.bot.clientService.BotService;
+import edu.java.bot.clientService.DefaultBotService;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
