+++ scrapper/src/main/java/edu/java/clientService/DefaultScrapperService.java
@@ -1,4 +1,4 @@
-package edu.java.services;
+package edu.java.clientService;
 
 import edu.java.api.httpClient.BotClient;
 import edu.java.dto.GenericResponse;
